requirements.txt
accelerate==0.23.0
aiohttp==3.8.6
aiosignal==1.3.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
autoflake==2.2.1
autoimport==1.3.3
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.10.0
bleach==6.1.0
blinker==1.6.3
cachetools==5.5.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.7
comm==0.1.4
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distro==1.9.0
docopt==0.6.2
docstring-to-markdown==0.13
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.104.0
fastjsonschema==2.18.1
filelock==3.12.4
fire==0.5.0
Flask==3.0.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gdown==5.2.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.0.0
grip==4.6.2
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.17.3
icdiff==2.0.7
idna==3.4
importlib-metadata==6.8.0
ipdb==0.13.13
ipykernel==6.25.2
ipython==8.16.1
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.17.2
Jinja2==3.1.2
jiter==0.6.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
lit==17.0.3
maison==1.4.0
Markdown==3.5
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
mypy==1.6.1
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
neovim==0.3.1
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1.1
notebook_shim==0.2.3
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
openai==1.52.2
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.7.1
path-and-address==2.0.1
pathspec==0.11.2
peft==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
pluggy==1.3.0
powerline-status==2.7
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.24.0
protobuf==5.28.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py3nvml==0.2.7
pyarrow==13.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==1.10.13
pyflakes==3.1.0
Pygments==2.16.1
pynvim==0.4.3
pyparsing==3.2.0
pyprojroot==0.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
ray==2.7.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.4
rsa==4.9
safetensors==0.4.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sh==2.0.6
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.34.1
triton==2.0.0
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.8.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.18.0
visidata==2.11.1
vllm==0.2.1.post1
watchfiles==0.21.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.0
xdg==6.0.0
xformers==0.0.22
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
